[FIX] eslint와 prettier 충돌 수정
