+++ eslint.config.js
@@ -1,44 +1,39 @@
 import js from "@eslint/js";
 import globals from "globals";
+import react from "eslint-plugin-react";
 import reactHooks from "eslint-plugin-react-hooks";
 import reactRefresh from "eslint-plugin-react-refresh";
-import unused_imports from "eslint-plugin-unused-imports";
-import react from "eslint-plugin-react";
+import unusedImports from "eslint-plugin-unused-imports";
 import prettier from "eslint-plugin-prettier";
+import prettierConfig from "eslint-config-prettier";
+
 export default [
-    {
-        settings: {
-            react: {
-                version: "detect",
-            },
-        },
-    },
-    { ignores: ["dist"] },
     {
         files: ["**/*.{js,jsx}"],
         languageOptions: {
             ecmaVersion: 2020,
-            globals: globals.browser,
+            sourceType: "module",
             parserOptions: {
-                ecmaVersion: "latest",
                 ecmaFeatures: { jsx: true },
-                sourceType: "module",
+            },
+            globals: {
+                ...globals.browser,
             },
         },
         plugins: {
+            react,
             "react-hooks": reactHooks,
             "react-refresh": reactRefresh,
-            "unused-imports": unused_imports,
-            react: react,
-            prettier: prettier,
+            "unused-imports": unusedImports,
+            prettier,
         },
         rules: {
             ...js.configs.recommended.rules,
             ...reactHooks.configs.recommended.rules,
-            "react-refresh/only-export-components": ["warn", { allowConstantExport: true }],
             "prettier/prettier": "error",
             "react/jsx-uses-react": "error",
             "react/jsx-uses-vars": "error",
+            "react-refresh/only-export-components": ["warn", { allowConstantExport: true }],
             "@typescript-eslint/no-unused-vars": "off",
             "unused-imports/no-unused-imports": "error",
             "unused-imports/no-unused-vars": [
@@ -52,4 +47,5 @@ export default [
             ],
         },
     },
+    prettierConfig,
 ];
