+++ package.json
@@ -23,6 +23,7 @@
     "react-router-dom": "^7.4.0"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.3.1",
     "@eslint/js": "^9.21.0",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
